requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.41.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
contexttimer==0.3.3
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
docstring_parser==0.16
einops==0.7.0
einops-exts==0.0.4
exceptiongroup==1.2.0
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.3
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gradio==3.35.2
gradio_client==0.2.9
h11==0.14.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.22.2
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
linkify-it-py==2.0.3
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
orjson==3.10.0
packaging==24.0
pandas==2.2.1
peft==0.4.0
pillow==10.3.0
protobuf==5.26.1
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.32.3
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.13.0
semantic-version==2.10.0
sentencepiece==0.1.99
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tenacity==8.4.2
threadpoolctl==3.4.0
timm==0.6.13
tokenizers==0.13.3
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.33.3
triton==2.2.0
trl==0.9.4
typing_extensions==4.10.0
tyro==0.8.5
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
websockets==12.0
xformers==0.0.25.post1
xxhash==3.4.1
yarl==1.9.4
